refactor(routers): extract shared HTTP dependencies
